LayoutTests/imported/w3c/web-platform-tests/css/css-view-transitions/style-inheritance-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_equals: group expected "rgb(255, 0, 0)" but got ""

Harness Error (FAIL), message = Unhandled rejection: assert_equals: group expected "rgb(255, 0, 0)" but got ""

PASS style inheritance of pseudo elements
